Move some React import files around
