+++ WcaOnRails/app/webpacker/components/SearchWidget/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import OmnisearchInput from './SearchWidget/OmnisearchInput';
-import { omnisearchApiUrl } from '../lib/requests/routes.js.erb';
+import OmnisearchInput from './OmnisearchInput';
+import { omnisearchApiUrl } from '../../lib/requests/routes.js.erb';
 
 const SearchWidget = () => (
   <OmnisearchInput
